part 4 fin

+++ Part4/controllers/blogs.js
@@ -1,6 +1,5 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
 
 blogsRouter.get('/', async (_request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
@@ -8,16 +7,13 @@ blogsRouter.get('/', async (_request, response) => {
 });
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body);
-
-  const users = await User.find({});
-  const user = users[0];
+  const user = request.user;
 
   if (!user) {
-    return response.status(400).json({ error: 'userId missing or not valid' });
+    return response.status(401).json({ error: 'token missing or invalid' });
   }
 
-  blog.user = user.id;
+  const blog = new Blog({ ...request.body, user: user.id });
 
   const savedBlog = await blog.save();
   user.blogs = user.blogs.concat(savedBlog._id);
@@ -27,6 +23,17 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = request.user;
+
+  if (!user) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized user' });
+  }
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
